Move the MakeOIDCRoleBinding() and MakeOIDCRole() to a separate file
